Added weapons gunAnimHelper, Sounds, animations, BulletCOntroller, BulletPrehub, CustomExeption. Now Pistol can shooting and reloading

+++ Assets/Scripts/Weapon.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public abstract class Weapon : MonoBehaviour
 {
-    [SerializeField] protected float damage;
-    [SerializeField] protected float range;
+    [SerializeField] protected int damage;
+    [SerializeField] protected int range;
     [SerializeField] protected float fireRate; // скорострельность
     protected float lastShotTime = int.MinValue; // время последнего выстрела
     public abstract void Attack();
